+++ doc/source/history.rst
@@ -25,6 +25,8 @@ Release history
 
 * ADDED: support for struct types as ``Computation`` arguments (for them, the ``ctypes`` attributes contain the corresponding module obtained with :py:func:`~reikna.cluda.dtypes.ctype_module`).
 
+* ADDED: support for non-sequential axes in :py:class:`~reikna.reduce.Reduce`.
+
 * FIXED: bug in the interactive ``Thread`` creation (reported by James Bergstra).
 
 * FIXED: Py3-incompatibility in the interactive ``Thread`` creation.